+++ django_tables2/__init__.py
@@ -22,7 +22,7 @@ from .paginators import LazyPaginator
 from .views import SingleTableMixin, SingleTableView, MultiTableMixin
 
 
-__version__ = "2.0.0b3"
+__version__ = "2.0.0b4"
 
 __all__ = (
     "Table",