+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## unreleased
+## 2.0.0b4 (2018-08-29)
  - The `klass` argument to `Table.paginate()` is renamed to `paginator_class`
  - Table views/mixins now take `ListView` attributes `paginator_class` and `paginate_orphans` into account.